requirements.txt
llama-parse
llama-index
paddlepaddle
paddleocr
pdf2image
numpy
Pillow
python-dotenv
streamlit
os